+++ cpp/include/dump.h
@@ -1,4 +1,5 @@
-#pragma once
+#ifndef DUMP_H_
+#define DUMP_H_
 
 #include <cstdint>
 #include <vector>
@@ -25,19 +26,30 @@ enum DumpStyle {
 
 /*
  * Base DumpSnapshot class
- *
  * */
-template <typename T>
 class DumpSnapshot {
  public:
   DumpSnapshot() = default;
-  DumpSnapshot(uint8_t dumpStyle, uint64_t timestep);
-  DumpSnapshot(uint8_t dumpStyle, uint64_t timestep, uint64_t natoms,
-               const SimulationBox &box);
-  DumpSnapshot(uint8_t dumpStyle, uint64_t timestep, uint64_t natoms,
-               const SimulationBox &box, const std::vector<Atom> &atoms);
 
-  ~DumpSnapshot();
+  DumpSnapshot(Int32 dumpStyle, Int64 timestep)
+      : dumpStyle_(dumpStyle), timestep_(timestep) {}
+
+  DumpSnapshot(Int32 dumpStyle, Int64 timestep, Int64 natoms,
+               const SimulationBox &box)
+      : dumpStyle_(dumpStyle),
+        timestep_(timestep),
+        natoms_(natoms),
+        box_(box) {}
+
+  DumpSnapshot(Int32 dumpStyle, Int64 timestep, Int64 natoms,
+               const SimulationBox &box, const std::vector<Atom> &atoms)
+      : dumpStyle_(dumpStyle),
+        timestep_(timestep),
+        natoms_(natoms),
+        box_(box),
+        atoms_(atoms) {}
+
+  //  ~DumpSnapshot();
 
   // Copy constructor
   DumpSnapshot(const DumpSnapshot &snapshot);
@@ -49,17 +61,42 @@ class DumpSnapshot {
   void append(const Atom &atom);
 
   // Subscript operator (read/write)
-  const Atom &operator[](uint64_t index) const;
-  Atom &operator[](uint64_t index);
+  const Atom &operator[](Int64 index) const;
+  Atom &operator[](Int64 index);
 
-  // Return a range/view of selected atoms
+  // Getters and setters
+  const auto &dumpStyle() const { return dumpStyle_; }
+  auto &dumpStyle() {
+    return const_cast<decltype(dumpStyle_) &>(
+        const_cast<const DumpSnapshot *>(this)->dumpStyle());
+  }
+
+  const auto &timestep() const { return timestep_; }
+  auto &timestep() {
+    return const_cast<decltype(timestep_) &>(
+        const_cast<const DumpSnapshot *>(this)->timestep());
+  }
+
+  const auto &natoms() const { return natoms_; }
+  auto &natoms() {
+    return const_cast<decltype(natoms_) &>(
+        const_cast<const DumpSnapshot *>(this)->natoms());
+  }
+
+  const auto &box() const { return box_; }
+  auto &box() {
+    return const_cast<decltype(box_) &>(
+        const_cast<const DumpSnapshot *>(this)->box());
+  }
 
  private:
-  uint8_t dumpStyle;
-  uint64_t timestep;
-  uint64_t natoms;
-  SimulationBox box;
-  std::vector<Atom> atoms;
+  Int32 dumpStyle_;
+  Int64 timestep_;
+  Int64 natoms_;
+  SimulationBox box_;
+  std::vector<Atom> atoms_;
 };
 
 }  // namespace lmptools
+
+#endif