+++ cpp/src/dump.cpp
@@ -1,0 +1,21 @@
+#include "dump.h"
+
+#include "simulation.h"
+
+/*lmptools::DumpSnapshot::DumpSnapshot(Int32 dumpStyle, Int64 timestep)
+    : dumpStyle_(dumpStyle), timestep_(timestep) {}
+
+lmptools::DumpSnapshot::DumpSnapshot(Int32 dumpStyle, Int64 timestep,
+                                     Int64 natoms, const SimulationBox& box)
+    : dumpStyle_(dumpStyle), timestep_(timestep), natoms_(natoms), box_(box) {}
+
+lmptools::DumpSnapshot::DumpSnapshot(Int32 dumpStyle, Int64 timestep,
+                                     Int64 natoms, const SimulationBox& box,
+                                     const std::vector<Atom>& atoms)
+    : dumpStyle_(dumpStyle),
+      timestep_(timestep),
+      natoms_(natoms),
+      box_(box),
+      atoms_(atoms) {}
+
+// lmptools::DumpSnapshot::~DumpSnapshot() {}*/