+++ tests/cpp/dump_unittest.cpp
@@ -1,3 +1,0 @@
-#include <gtest/gtest.h>
-
-TEST(Pass, Always) { EXPECT_EQ(1, 1); }